Ajuste: - Refactorizacion de mensajes estaticos y el manejo de excepciones

+++ servicio-clientes/src/main/java/com/example/pragma/servicioclientes/infrastructure/entrypoints/RecursoClientes.java
@@ -10,8 +10,9 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
-import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -19,14 +20,22 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.util.List;
 
+import static com.example.pragma.servicioclientes.application.config.messages.constants.ApiErrorResponses.CLIENTE_NO_ENCONTRADO;
+import static com.example.pragma.servicioclientes.application.config.messages.constants.ApiErrorResponses.CLIENTE_YA_EXISTE;
+
 @RestController
 @RequestMapping("/clientes")
+@RequiredArgsConstructor
 public class RecursoClientes {
 
     @Autowired
     private GestionarClientesInterface gestionarClientes;
+
     @Autowired
     private MapeadorDtoModelo mapeadorDto;
+
+    @Autowired
+    private MessageSource messageSource;
 
     @ApiOperation("Colsultar el cliente por identificacion")
     @ApiResponses({
@@ -42,7 +51,10 @@ public class RecursoClientes {
     ) {
         Cliente clienteBD = gestionarClientes.consultarCliente(tipoIdentificacion, numeroIdentificacion);
         if (clienteBD == null) {
-            return ResponseEntity.notFound().build();
+            throw new ApiException(
+                    HttpStatus.NOT_FOUND,
+                    messageSource.getMessage(CLIENTE_NO_ENCONTRADO, null, null)
+            );
         }
         return ResponseEntity.ok(mapeadorDto.aRespuesta(clienteBD));
     }
@@ -70,9 +82,12 @@ public class RecursoClientes {
             @Valid
             @ApiParam(value = "cliente en formato Json", required = true)
             @RequestBody ClientePeticion clientePeticion
-            ) {
+    ) {
         if (existeElCliente(clientePeticion)) {
-            throw new ApiException(HttpStatus.BAD_REQUEST, "El cliente ya existe");
+            throw new ApiException(
+                    HttpStatus.BAD_REQUEST,
+                    messageSource.getMessage(CLIENTE_YA_EXISTE, null, null)
+            );
         }
         Cliente cliente = mapeadorDto.aModelo(clientePeticion);
         Cliente clienteCreado = gestionarClientes.guardarCliente(cliente);
@@ -91,6 +106,12 @@ public class RecursoClientes {
             @RequestBody ClientePeticion clientePeticion) {
 
         Cliente cliente = mapeadorDto.aModelo(clientePeticion);
+
+        if (!existeElCliente(clientePeticion)) {
+            throw new ApiException(
+                    HttpStatus.NOT_FOUND,
+                    messageSource.getMessage(CLIENTE_NO_ENCONTRADO, null, null));
+        }
         Cliente imagenActualizada = gestionarClientes.actualizarCliente(cliente);
         ClienteRespuesta respuesta = mapeadorDto.aRespuesta(imagenActualizada);
         return ResponseEntity.status(HttpStatus.OK).body(respuesta);
@@ -109,6 +130,14 @@ public class RecursoClientes {
             @ApiParam(value = "Identificador del cliente", required = true)
             @PathVariable String identificacion) {
 
+        Cliente clienteExistente = gestionarClientes.consultarCliente(
+                tipo,
+                identificacion);
+        if (clienteExistente == null) {
+            throw new ApiException(
+                    HttpStatus.NOT_FOUND,
+                    messageSource.getMessage(CLIENTE_NO_ENCONTRADO, null, null));
+        }
         gestionarClientes.eliminarCliente(tipo, identificacion);
         return ResponseEntity.ok().build();
     }
